Made changes to server.kt to make it a suspended function, able to reuse the same port once the server has closed. Added some extra exception handling for the cancellation of the server and for the possible socket errors.
Added a Run Server button to Main.kt. Made it able to toggle between running and closing the server, maintaining the state of the button and server object even when navigating outside of the landing page. Added some logic and delay to change the text and status of the button.

Empty space changes to client.kt.

TODO: Input Validation. Multi-line validation. OverviewPage text/description. Styling.